Fix computed properties

+++ app/guid-node/iqbrims/file-browser.ts
@@ -4,7 +4,6 @@ import { later } from '@ember/runloop';
 import { all, task, timeout } from 'ember-concurrency';
 
 import File from 'ember-osf-web/models/file';
-import FileProviderModel from 'ember-osf-web/models/file-provider';
 import Node from 'ember-osf-web/models/node';
 
 import GuidNodeIQBRIMS from './controller';
@@ -12,8 +11,8 @@ import GuidNodeIQBRIMS from './controller';
 export default class IQBRIMSFileBrowser extends EmberObject {
     static readonly FILES_TXT = '.files.txt';
 
-    folderName: string;
-    owner: GuidNodeIQBRIMS;
+    folderName: string | null = null;
+    owner: GuidNodeIQBRIMS | null = null;
 
     filter: string = this.filter || '';
     sort: string = this.sort || 'name';
@@ -59,7 +58,7 @@ export default class IQBRIMSFileBrowser extends EmberObject {
 
     addFile = task(function *(this: IQBRIMSFileBrowser, id: string) {
         const allFiles = this.get('allFiles');
-        if (!allFiles) {
+        if (!allFiles || !this.owner) {
             return;
         }
         const duplicate = allFiles.findBy('id', id);
@@ -84,6 +83,9 @@ export default class IQBRIMSFileBrowser extends EmberObject {
     });
 
     deleteFile = task(function *(this: IQBRIMSFileBrowser, file: File) {
+        if (!this.owner) {
+            return;
+        }
         try {
             yield file.destroyRecord();
             yield this.get('flash').perform(file, this.owner.get('intl').t('file_browser.file_deleted'), 'danger');
@@ -105,6 +107,9 @@ export default class IQBRIMSFileBrowser extends EmberObject {
     });
 
     moveFile = task(function *(this: IQBRIMSFileBrowser, file: File, node: Node): IterableIterator<any> {
+        if (!this.owner) {
+            return;
+        }
         try {
             yield file.move(node);
             yield this.get('flash').perform(file, this.owner.get('intl').t('file_browser.successfully_moved'));
@@ -126,6 +131,9 @@ export default class IQBRIMSFileBrowser extends EmberObject {
         conflict?: string,
         conflictingFile?: File,
     ) {
+        if (!this.owner) {
+            return;
+        }
         const flash = this.get('flash');
 
         try {
@@ -148,41 +156,32 @@ export default class IQBRIMSFileBrowser extends EmberObject {
         }
     });
 
-    constructor(owner: GuidNodeIQBRIMS, folderName: string) {
-        super();
-        this.owner = owner;
-        this.folderName = folderName;
-    }
-
     @computed('owner.workingDirectory.files.[]')
     get targetDirectory(): File | undefined {
-        if (!this.owner.workingDirectory) {
+        if (!this.owner) {
             return undefined;
         }
-        return this.findTargetDirectory(this.owner.workingDirectory);
-    }
-
-    findTargetDirectory(defaultStorage: File) {
-        if (!defaultStorage.files.isFulfilled && !defaultStorage.files.isRejected) {
-            later(() => {
-                this.findTargetDirectory(defaultStorage);
-            }, 500);
+        const defaultStorage = this.owner.get('workingDirectory');
+        if (!defaultStorage) {
             return undefined;
         }
-        const files = defaultStorage.files.filter(f => f.name === this.folderName);
+        const storageFiles = defaultStorage.get('files');
+        const files = storageFiles.filter(f => f.name === this.folderName);
         if (files.length === 0) {
             this.createTargetDirectory(defaultStorage);
             return undefined;
         }
-        this.notifyPropertyChange('allFiles');
         return files[0];
     }
 
-    createTargetDirectory(defaultStorage: File) {
-        if (this.newFolderRequest) {
+    createTargetDirectory(defaultStorage: File | undefined) {
+        if (!defaultStorage) {
             return;
         }
-        const newFolderUrl = defaultStorage.links.new_folder;
+        if (this.newFolderRequest || !this.owner || !this.folderName) {
+            return;
+        }
+        const newFolderUrl = defaultStorage.get('links').new_folder;
         this.newFolderRequest = this.owner.currentUser.authenticatedAJAX({
             url: `${newFolderUrl}&name=${encodeURIComponent(this.folderName)}`,
             type: 'PUT',
@@ -215,31 +214,20 @@ export default class IQBRIMSFileBrowser extends EmberObject {
         return undefined;
     }
 
-    @computed('owner.gdProvider')
+    @computed('owner.gdProvider.files.[]')
     get gdTargetDirectory(): File | undefined | null {
-        if (this.owner.gdProvider === undefined) {
+        if (!this.owner) {
             return undefined;
         }
-        if (this.owner.gdProvider === null) {
-            this.set('gdLoading', false);
-            return null;
-        }
-        return this.findGDTargetDirectory(this.owner.gdProvider);
-    }
-
-    findGDTargetDirectory(defaultStorage: FileProviderModel) {
-        if (!defaultStorage.rootFolder.isFulfilled && !defaultStorage.rootFolder.isRejected) {
-            later(() => {
-                this.findGDTargetDirectory(defaultStorage);
-                this.notifyPropertyChange('gdTargetDirectory');
-            }, 500);
+        const gdProvider = this.owner.get('gdProvider');
+        if (!gdProvider) {
             return undefined;
         }
-        const files = defaultStorage.rootFolder.files.filter(f => f.name === this.folderName);
+        const storageFiles = gdProvider.get('files');
+        const files = storageFiles.filter(f => f.name === this.folderName);
         if (files.length === 0) {
             return undefined;
         }
-        this.notifyPropertyChange('gdTargetDirectory');
         return files[0];
     }
 
@@ -295,6 +283,9 @@ export default class IQBRIMSFileBrowser extends EmberObject {
 
     @action
     async openFile(this: IQBRIMSFileBrowser, file: File, show: string) {
+        if (!this.owner) {
+            return;
+        }
         const guid = file.get('guid') || await file.getGuid();
 
         this.owner.transitionToRoute('guid-file', guid, { queryParams: { show } });
